requirements.txt
azure-functions==1.18.0
python-dotenv==1.0.1
